requirements.txt
einops==0.7.0
torch>=1.7
basicsr==1.3.4.9
pyyaml==6.0.1
python-box==6.0.2
ensure==1.0.2   
pathlib==1.0.1
wget
tqdm==4.65.0
opencv-python==4.8.1.78
numpy